Move export code to special class
